Restructure the 'skills' folder for clarity.
